pyproject.toml
[project]
name = "vibe-journey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]>=0.201.1",
    "flask>=3.1.2",
    "ipykernel>=6.30.1",
    "python-dotenv>=1.1.1",
]
